refactor: centralize offline status update
